Update per-channel to include outside gamut prism values
Update the per-channel handling to include values
that exceed the gamut. This requires a bit of a
trick to pull off because by default, all values
must evaluate to the aesthetic transfer function curve
limits, thereby clipping them at the limits of the
curve. To facilitate negative values relative to the
RGB, we absolute the values, evaluate the curve, then
return the negative sign.